Moved files into more useful folder system.
