+++ idk/src/main/java/com/treydev/idk/attack/CombatHandler.java
@@ -1,6 +1,10 @@
-package com.treydev.idk;
+package com.treydev.idk.attack;
 
 import java.util.ArrayList;
+
+import com.treydev.idk.IdkController;
+import com.treydev.idk.critter.Critter;
+import com.treydev.idk.critter.Species;
 
 public class CombatHandler {
     public ArrayList<String> outputs;
