+++ idk/src/main/java/com/treydev/idk/attack/CombatModel.java
@@ -1,4 +1,4 @@
-package com.treydev.idk;
+package com.treydev.idk.attack;
 
 import java.util.ArrayList;
