+++ idk/src/main/java/com/treydev/idk/attack/Move.java
@@ -1,8 +1,12 @@
-package com.treydev.idk;
+package com.treydev.idk.attack;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Random;
+
+import com.treydev.idk.Element;
+import com.treydev.idk.SpecialAttack;
+import com.treydev.idk.critter.Critter;
 
 public abstract class Move {
