+++ idk/src/main/java/com/treydev/idk/critter/CritterDetailsModel.java
@@ -1,4 +1,4 @@
-package com.treydev.idk;
+package com.treydev.idk.critter;
 
 public class CritterDetailsModel {
     public class statLine {
