+++ idk/src/test/java/com/treydev/idk/MoveTests.java
@@ -6,6 +6,9 @@ import java.util.HashMap;
 import org.junit.jupiter.api.Test;
 import org.springframework.util.Assert;
 
+import com.treydev.idk.attack.Move;
+import com.treydev.idk.critter.Critter;
+
 public class MoveTests {
     @Test
     void testGenMoveList() {
